Rescale covariates to have unit variance
Now rescales covariate vector such that each predictor variable has variance 1. This improves optimization performance in datasets where there are large differences in the magnitude of predictors.

Still needs to be modified to work well for DiffSynth, but it constitutes a major improvement to the Synth implementation and thus warrants a new push.