Leaving this for now - might implement in future
Reason: Special predictors allow for imho too easy specification-searching.
If you select ways to average covariates, run the model, look at the comparison df and the validity tests etc, and then decide to change specifcation because you were unhappy with the results–you are p-hacking. The comparison df and validity tests are the means by which significance of results are evalauted in Synthetic Controls. Changing model based on these is equivalent to p-hacking through specification-search.

It also begs a lot of questions about what is the correct way to perform in-time placeboes. For example, let's say
the treatment period is 1990 and I include a special predictor that is the outcome in 1980. If I then run an in-time placebo starting 1980 – what should I do with the special predictor? Should I move it back 10 years before the placebo treatment? What if the first observation in from 1971?